+++ example/utils/PathTracingViewer.js
@@ -3,10 +3,7 @@ import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls.js';
 import { FullScreenQuad } from 'three/examples/jsm/postprocessing/Pass.js';
 import { SAH } from 'three-mesh-bvh';
 import { GenerateMeshBVHWorker } from 'three-mesh-bvh/src/workers/GenerateMeshBVHWorker.js';
-import { PathTracingRenderer } from '../core/PathTracingRenderer.js';
-import { mergeMeshes } from '../utils/GeometryPreparationUtils.js';
-import { PhysicalPathTracingMaterial } from '../materials/PhysicalPathTracingMaterial.js';
-import { MaterialReducer } from '../core/MaterialReducer.js';
+import { PathTracingRenderer, mergeMeshes, PhysicalPathTracingMaterial, MaterialReducer } from '../../src/index.js';
 
 export class PathTracingViewer {
